refactor: createCmsParagraph
