+++ src/actions/cms/articleSections/update.ts
@@ -4,7 +4,7 @@ import { destroyArticleSection } from './destroy'
 import prisma from '@/prisma'
 import { createActionError, createPrismaActionError } from '@/actions/error'
 import { createCmsImage } from '@/server/cms/images/create'
-import { createCmsParagraph } from '@/actions/cms/paragraphs/create'
+import { createCmsParagraph } from '@/server/cms/paragraphs/create'
 import { createCmsLink } from '@/actions/cms/links/create'
 import type { ImageSize, ArticleSection, Position } from '@prisma/client'
 import type { ExpandedArticleSection } from './Types'
